% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedPlot.R
\name{nestedPlot}
\alias{nestedPlot}
\title{Nested Plot}
\usage{
nestedPlot(
  df,
  v1,
  v2,
  v3,
  obs.y,
  iscat = T,
  freqz = F,
  pc = F,
  jW = 0.35,
  jH = 0.35,
  NoCls = F,
  limits = NULL,
  plot.Trend = F,
  method = "auto"
)
}
\arguments{
\item{df}{Data Frame Object}

\item{v1}{First variable (X-var of plot) that can be categorical or continuous.}

\item{v2}{Second variable (Y-var of plot) that can be categorical or continuous.}

\item{v3}{Third variable used to facet/group the X/Y variables, must be categorical.}

\item{obs.y}{Class variable of intrest, must be categorical. Can assigned NULL if NoCls = F.}

\item{iscat}{If X & Y are categorical, then this requires to be TRUE. If false scatter plots are used for nested plots.}

\item{freqz}{If iscat = TRUE, freqz displays frequency plots for nested categorys.}

\item{pc}{For "freqz" plots, adjusts values to display percentages.}

\item{jW}{Jitter width for numerical scatter plots.}

\item{jH}{Jitter height for numerical scatter plots.}

\item{NoCls}{Refrain from including classification variable.}

\item{limits}{For scatter plots input y limits.}
}
\value{
GGplot Object
}
\description{
This program creats nested plots for quick exploratory data analysis.
}
